Move NewPlatingItem(), GetLastePlateOrNew() from printer to HistoryContainerBase
